+++ src/main/java/com/exercise/basic/array/Bai6.java
@@ -1,10 +1,14 @@
-package com.exercise.basic;
+package com.exercise.basic.array;
 
 public class Bai6 {
 
     public int searchBinary(final int arr[], final int value) {
+        if (arr.length == 0) {
+            return -1;
+        }
+
         int first = 0;
-        int last = arr[arr.length - 1];
+        int last = arr.length - 1;
 
         while (first <= last) {
             final int mid = (first + last) / 2;
